+++ civis/base.py
@@ -4,6 +4,7 @@ import threading
 from concurrent import futures
 import warnings
 from requests import Request
+from json.decoder import JSONDecodeError
 
 from civis.response import PaginatedResponse, convert_response_data_type
 from civis._utils import open_session, retry_request, MAX_RETRIES
@@ -56,8 +57,17 @@ def _err_msg_with_job_run_ids(err_msg, job_id, run_id) -> str:
 class CivisAPIError(Exception):
     def __init__(self, response):
         if response.content:  # the API itself gave an error response
-            json = response.json()
-            self.error_message = json["errorDescription"]
+            try:
+                json = response.json()
+            except JSONDecodeError as e:
+                if "Expecting value: line 1 column 1 (char 0)" in str(e):
+                    self.error_message = "No Response Content from Civis API"
+                else:
+                    self.error_message = f"Response Content: " \
+                                    f"{(response.content or b'').decode()}"
+            else:
+                self.error_message = json.get("errorDescription") or \
+                                     "No Error Message Available"
         else:  # this was something like a 502
             self.error_message = response.reason